Renamed Algae folder to algaePivot
